Monthly love reset (#111)
* Made datetime import/usage nicer

* Added PyCharm Runner

* Updated rules and roles strings to not be ridiculously laid out

* Added Vars to config for saving variables.

* Updated Vars to supply default values

* Updated cfg to make db name not required

It can be automatically supplied; defaults to 'aryas'

* Further changes to layout of string constants

* Removed type hint from Vars constructor

This threw an exception.

* Renamed vars filename

The underlying implementation meant there were unnecessary file extentions tacked on.

* Added in_bot_channel() to utility.py

* Made bot info channel configurable.

The General class is now defined inside the setup function because dependancy injection is just so much fun.

* Added monthly_allowance to config schema

* Made bot info channel configurable.

The General class is now defined inside the setup function because dependancy injection is just so much fun.

* Changed User's total_love to a list comprehension

* Added next_love_reset property to Vars

* Added remaining_givable_love property to User model

* Commented out on_member_join

PEP8 threw a warning, wanted it to be static

* Small reformats in general.py

* Added check to show_love() to make sure user has enough ❤ left

* Updated get_love() to show remaining givable love

Also changed to use new f-strings. Shiny!

* Added reset check

The check interval is configurable; default is 10 minutes.